backend/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-auto-rest==0.1.7
django-seed==0.2.2
djangorestframework==3.12.2
drf-renderer-xlsx==0.3.9
et-xmlfile==1.0.1
Faker==6.5.0
idna==2.10
jdcal==1.4.1
numpy==1.20.1
openpyxl==3.0.6
pandas==1.2.3
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.3
